Move test to a file
